+++ components/studentAreas.js
@@ -1,0 +1,10 @@
+import renderToDOM from '../utils/renderToDOM';
+
+const studentAreas = () => {
+  const domString = `<div id="students">No Students</div>
+  <div id="voldy">No Death Eaters</div>`;
+
+  renderToDOM('#student-container', domString);
+};
+
+export default studentAreas;
