+++ utils/data/students.js
@@ -1,0 +1,27 @@
+import createId from '../createId';
+import studentsOnDom from '../../components/studentsOnDom';
+import houses from './houses';
+
+const students = [];
+
+const sortStudent = (e) => {
+  e.preventDefault();
+  const sortingHat = houses[Math.floor(Math.random() * houses.length)];
+
+  if (e.target.id === 'sorting') {
+    const student = document.querySelector('#student-name');
+
+    // create the new student object
+    students.push({
+      id: createId(students),
+      name: student.value,
+      house: sortingHat.house,
+      crest: sortingHat.crest
+    });
+
+    student.value = ''; // reset value of input
+    studentsOnDom('#students', students);
+  }
+};
+
+export { students, sortStudent };
